Refactored cmd/dashboard/{instancesocket,logstreamer}.go into dashboard.go so that it works. Changed log.Trace to log.Error when appropriate.
