+++ cmd/dashboard/dashboard.go
@@ -7,9 +7,13 @@ import (
 	"fmt"
 	"github.com/bketelsen/skynet"
 	"github.com/bketelsen/skynet/client"
+	"html"
 	"html/template"
 	"net/http"
 	"os"
+	"regexp"
+	"strings"
+	"time"
 )
 
 import (
@@ -43,7 +47,7 @@ func searchHandler(w http.ResponseWriter, r *http.Request) {
 	if session == nil {
 		session, err = mgo.Dial(*mgoserver)
 		if err != nil {
-			log.Trace(fmt.Sprintf("%+v", skynet.MongoError{
+			log.Error(fmt.Sprintf("%+v", skynet.MongoError{
 				*mgoserver, "can't connect to server",
 			}))
 			// Tell client:
@@ -61,7 +65,7 @@ func searchHandler(w http.ResponseWriter, r *http.Request) {
 	} else {
 		dbs, err = session.DatabaseNames()
 		if err != nil {
-			log.Trace(fmt.Sprintf("%+v", skynet.MongoError{
+			log.Error(fmt.Sprintf("%+v", skynet.MongoError{
 				*mgoserver,
 				fmt.Sprintf("unable to obtain database list: %s", err),
 			}))
@@ -179,4 +183,307 @@ func Doozer() *skynet.DoozerConnection {
 	conn.Connect()
 
 	return conn
+}
+
+//
+// Beginning of what was logstreamer.go
+//
+
+type connection struct {
+	ws     *websocket.Conn
+	db     string // mongodb database currently in use
+	coll   string // mongodb collection currently tailing
+	cancel chan bool
+	sess   *mgo.Session
+	filter *regexp.Regexp // if set, the client only sees matches
+}
+
+type Req struct {
+	Collection string `json:"collection"`
+	Filter     string `json:"filter"`
+}
+
+func (c *connection) send(s string) bool {
+	err := websocket.Message.Send(c.ws, s)
+	if err != nil {
+		c.ws.Close()
+		return false
+	}
+	return true
+}
+
+func (c *connection) fromClient() {
+	shouldCancel := false
+	message := &Req{}
+	for {
+		err := websocket.JSON.Receive(c.ws, message)
+		if err != nil {
+			if *debug {
+				fmt.Printf("%s: error receiving from client: %s\n",
+					c.ws.Request().RemoteAddr, err)
+			}
+			break
+		}
+		if *debug {
+			fmt.Printf("%s: fromClient: %+v\n", c.ws.Request().RemoteAddr, message)
+		}
+
+		if shouldCancel {
+			c.cancel <- true
+			shouldCancel = false
+		}
+
+		if message.Filter != "" {
+			c.filter, err = regexp.Compile(html.UnescapeString(message.Filter))
+			if err != nil {
+				s := fmt.Sprintf("reader: can not compile regexp: %s %s",
+					message.Filter, err)
+				if !c.send(s) {
+					return
+				}
+				if *debug {
+					fmt.Printf("%s: %s\n", c.ws.Request().RemoteAddr, s)
+				}
+				continue
+			}
+		} else {
+			c.filter = nil
+		}
+
+		if message.Collection != "" {
+			dbc := strings.Split(message.Collection, ":")
+			if len(dbc) != 2 {
+				s := fmt.Sprintf("internal error: received bad db:collection from client: %s", message.Collection)
+				if !c.send(s) {
+					return
+				}
+				if *debug {
+					fmt.Printf("%s: %s\n", c.ws.Request().RemoteAddr, s)
+				}
+				continue
+			}
+			c.db = dbc[0]
+			c.coll = dbc[1]
+		} else {
+			s := fmt.Sprintf("internal error: db:collection shouldn't be nil")
+			if !c.send(s) {
+				return
+			}
+			if *debug {
+				fmt.Printf("%s: %s\n", c.ws.Request().RemoteAddr, s)
+			}
+			c.db = ""
+			c.coll = ""
+			continue
+		}
+
+		if c.coll != "" {
+			go c.dump()
+			shouldCancel = true
+		}
+	}
+	c.ws.Close()
+}
+
+func (c *connection) dump() {
+	var result = make(map[string]interface{})
+
+	coll := c.sess.DB(c.db).C(c.coll)
+	if coll == nil {
+		if !c.send("internal mgo error: shouldn't happen!") {
+			return
+		}
+		<-c.cancel
+		return
+	}
+	iter := coll.Find(nil).Tail(500 * time.Millisecond)
+	if iter.Err() != nil {
+		s := fmt.Sprintf("internal error: %s", iter.Err())
+		if !c.send(s) {
+			return
+		}
+		if *debug {
+			fmt.Printf("%s: %s\n", c.ws.Request().RemoteAddr, s)
+		}
+		// we must block here, no need to continue spinning
+		<-c.cancel
+		return
+	}
+
+	// Need to spin to be able to consume a cancel
+	for {
+		select {
+		case <-c.cancel:
+			return
+		default:
+			if iter.Next(result) {
+				for k, v := range result {
+					// how complicated are the logs? objects? collections?
+					s := fmt.Sprintf("%v: %v", k, v)
+					if c.filter == nil || c.filter.MatchString(s) {
+						if !c.send(s) {
+							return
+						}
+					}
+				}
+				if iter.Err() != nil {
+					s := fmt.Sprintf("%s", iter.Err())
+					if !c.send(s) {
+						return
+					}
+					<-c.cancel
+					return
+				}
+			} else {
+				if iter.Err() != nil {
+					s := fmt.Sprintf("%s", iter.Err())
+					if !c.send(s) {
+						return
+					}
+					<-c.cancel
+					return
+				}
+				if !iter.Timeout() {
+					if !c.send("lost connection to server, won't retry") {
+						return
+					}
+					<-c.cancel
+					return
+				}
+			}
+		}
+	}
+}
+
+func wsHandler(ws *websocket.Conn) {
+	// Would it be better for each individual client to open 
+	// a separate connection with the MongoDB server by calling Dial here?
+	c := &connection{ws: ws, cancel: make(chan bool), sess: session}
+	c.fromClient() // must wait for client to select database
+	ws.Close()
+	close(c.cancel) // ensure no dangling readers
+}
+
+//
+// Beginning of what was instancesocket.go
+//
+
+type SocketResponse struct {
+	Action string
+	Data   interface{}
+}
+
+type SocketRequest struct {
+	Action string
+	Data   interface{}
+}
+
+func instanceSocketRead(ws *websocket.Conn, readChan chan SocketRequest, closeChan chan bool) {
+	// Watch for read, if it fails break out of loop and close
+	for {
+		var request SocketRequest
+		err := websocket.JSON.Receive(ws, &request)
+
+		if err != nil {
+			closeChan <- true
+			break
+		}
+
+		readChan <- request
+	}
+}
+
+func sendInstanceList(ws *websocket.Conn, im *client.InstanceMonitor) {
+	// Wait for list to be built, then pull off the notification channel
+}
+
+func NewInstanceSocket(ws *websocket.Conn, im *client.InstanceMonitor) {
+	closeChan := make(chan bool, 1)
+	readChan := make(chan SocketRequest)
+	ticker := time.NewTicker(5 * time.Second)
+	lastHeartbeat := time.Now()
+
+	go instanceSocketRead(ws, readChan, closeChan)
+
+	l := im.Listen(skynet.UUID(), &skynet.Query{}, true)
+
+	instances := <-l.NotificationChan
+	err := websocket.JSON.Send(ws, SocketResponse{Action: "List", Data: instances})
+
+	if err != nil {
+		closeChan <- true
+	}
+
+	for {
+		select {
+		case <-closeChan:
+			ticker.Stop()
+			ws.Close()
+			l.Close()
+		case t := <-ticker.C:
+			// Check for timeout
+			if t.Sub(lastHeartbeat) > (15 * time.Second) {
+				closeChan <- true
+			}
+		case request := <-readChan:
+			lastHeartbeat = time.Now()
+
+			switch request.Action {
+			case "List":
+				err := websocket.JSON.Send(ws, SocketResponse{Action: "List", Data: instances})
+				if err != nil {
+					closeChan <- true
+				}
+			case "Heartbeat":
+				// this is here more for documentation purposes,
+				// setting the lastHeartbeat on read handles the logic
+				// here
+			case "Filter":
+				if request.Data != nil {
+					data := request.Data.(map[string]interface{})
+
+					if r, ok := data["Reset"]; ok {
+						reset := r.(bool)
+						if reset {
+							l.Query.Reset()
+						}
+					}
+
+					if r, ok := data["Registered"]; ok {
+						filter := r.(bool)
+						l.Query.Registered = &filter
+					}
+				}
+
+				instances := l.GetInstances()
+				iln := make(client.InstanceListenerNotification)
+				for _, i := range instances {
+					path := i.GetConfigPath()
+					iln[path] = client.InstanceMonitorNotification{
+						Path:    path,
+						Service: i,
+						Type:    client.InstanceAddNotification,
+					}
+				}
+
+				err := websocket.JSON.Send(ws, SocketResponse{Action: "List", Data: iln})
+
+				if err != nil {
+					closeChan <- true
+				}
+			}
+
+		case notification := <-l.NotificationChan:
+			var err error
+
+			// Forward message as it stands across the websocket
+			err = websocket.JSON.Send(ws, SocketResponse{Action: "Update", Data: notification})
+
+			instances = instances.Join(notification)
+
+			if err != nil {
+				closeChan <- true
+			}
+		}
+	}
 }
